Refactor the file structure to follow bulletproof-react
