+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "ryt-bank-assessment",
   "version": "1.0.0",
-  "main": "index.ts",
+  "main": "src/app/index.ts",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
