+++ src/app/app.json
