+++ src/components/Card/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { View, StyleSheet, ViewStyle } from 'react-native';
-import { COLORS, SPACING } from '../../constants';
+import { COLORS, SPACING } from '../constants';
 
 interface CardProps {
   children: React.ReactNode;
